pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "A boilerplate FastAPI project with Uvicorn for serving APIs"
authors = ["Viktor <limphanith.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
sqlalchemy = "^2.0.34"
fastapi = "0.111.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"